refactor(migrations): consolidate chunking migrations into single file
- Merged 11 conflicting chunking migrations into one comprehensive migration
- Removed duplicate operations and conflicting table creations
- Added idempotent operations for partial state handling
- Ensured both upgrade and downgrade paths work correctly
- Uses optimal 100 LIST partitions from the start
- Includes all indexes, views, and monitoring functions
- Tested with linting and unit tests

This consolidation eliminates migration conflicts and simplifies the
schema evolution for the pre-release codebase.

Fixes TICKET-009
